Updated README to make note of existing Neo field issue
Thought it might be helpful to alert users to lack of Neo support, before they download and install it. 

I know that for our agency, Neo is one of the core plugins we almost always use, and it's also one of the key reasons we were looking for a way to import/export fields because they can take so long to setup. :-P